backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-nvidia-ai-endpoints>=0.2.0
langchain-openai>=0.2.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1

# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
